+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Add reason codes to the `X-Sentry-Rate-Limits` header in store responses. This allows external Relays to emit outcomes with the proper reason codes. ([#850](https://github.com/getsentry/relay/pull/850))
 - Emit metrics for outcomes in external relays. ([#851](https://github.com/getsentry/relay/pull/851))
 - Make `$error.value` `pii=true`. ([#837](https://github.com/getsentry/relay/pull/837))
+- Send `key_id` in partial project config. ([#854](https://github.com/getsentry/relay/pull/854))
 
 ## 20.11.1